fix(boot): Live/Install tick types; QEMU OTA 4G smp1 (ADR-0100 T-017 T-046 T-047)
USB unified image built locally (usb_hw.img not in git). Host OTA GET/POST 200 in docs/evidence. Guest e2e A2 still open (L3_5_FAIL).

+++ crates/neural-kernel/src/agents.rs
@@ -193,7 +193,7 @@ impl InputAgent {
         if scancode >= 0x80 { return; }
         // ADR-0100 T-028: ~5s pós-boot, [L]ive / [I]nstall (default Live já é T-029).
         {
-            let ticks = crate::interrupts::TIMER_TICKS.load(core::sync::atomic::Ordering::Relaxed);
+            let ticks = crate::interrupts::TIMER_TICKS.load(core::sync::atomic::Ordering::Relaxed) as u64;
             let hz = crate::interrupts::TIMER_HZ.load(core::sync::atomic::Ordering::Relaxed).max(1);
             if ticks < hz.saturating_mul(5) {
                 match scancode {
